Moves some logic out of the mod.rs
Makes a start on #61
