+++ pb-commit-message-lints/src/lints/mod.rs
@@ -97,36 +97,6 @@ impl std::convert::From<Lint> for String {
     }
 }
 
-/// Get the lints that are currently enabled
-///
-/// # Errors
-///
-/// If there's an error reading from the configuration source
-pub fn get_lint_configuration(config: &dyn Vcs) -> Result<Lints, PbCommitMessageLintsError> {
-    Ok(Lints::new(
-        vec![
-            get_config_or_default(config, Lint::DuplicatedTrailers, true)?,
-            get_config_or_default(config, Lint::PivotalTrackerIdMissing, false)?,
-            get_config_or_default(config, Lint::JiraIssueKeyMissing, false)?,
-        ]
-        .into_iter()
-        .flatten()
-        .collect(),
-    ))
-}
-
-fn get_config_or_default(
-    config: &dyn Vcs,
-    lint: Lint,
-    default: bool,
-) -> Result<Option<Lint>, PbCommitMessageLintsError> {
-    Ok(config
-        .get_bool(&lint.config_key())?
-        .or(Some(default))
-        .filter(|lint_value| lint_value == &true)
-        .map(|_| lint))
-}
-
 #[cfg(test)]
 mod tests_lints {
     use std::convert::TryInto;
@@ -161,121 +131,6 @@ mod missing_pivotal_tracker_id;
 mod duplicate_trailers;
 
 mod missing_jira_issue_key;
-
-#[cfg(test)]
-mod tests_get_lint_configuration {
-    use std::collections::BTreeMap;
-
-    use pretty_assertions::assert_eq;
-
-    use crate::lints::lib::Lints;
-    use crate::{
-        errors::PbCommitMessageLintsError,
-        external::vcs::InMemory,
-        lints::{
-            get_lint_configuration,
-            Lint::{DuplicatedTrailers, JiraIssueKeyMissing, PivotalTrackerIdMissing},
-        },
-    };
-
-    #[test]
-    fn defaults() {
-        let mut strings = BTreeMap::new();
-        let config = InMemory::new(&mut strings);
-
-        let actual = get_lint_configuration(&config);
-        let expected = Ok(Lints::new(vec![DuplicatedTrailers]));
-
-        assert_eq!(
-            expected, actual,
-            "Expected the list of lint identifiers to be {:?}, instead got {:?}",
-            expected, actual
-        )
-    }
-
-    #[test]
-    fn disabled_duplicated_trailers() {
-        let mut strings = BTreeMap::new();
-        strings.insert("pb.lint.duplicated-trailers".into(), "false".into());
-        let config = InMemory::new(&mut strings);
-
-        let actual = get_lint_configuration(&config);
-        let expected: Result<Lints, PbCommitMessageLintsError> = Ok(Lints::new(vec![]));
-
-        assert_eq!(
-            expected, actual,
-            "Expected the list of lint identifiers to be {:?}, instead got {:?}",
-            expected, actual
-        )
-    }
-
-    #[test]
-    fn enabled_duplicated_trailers() {
-        let mut strings = BTreeMap::new();
-        strings.insert("pb.lint.duplicated-trailers".into(), "true".into());
-        let config = InMemory::new(&mut strings);
-
-        let actual = get_lint_configuration(&config);
-        let expected = Ok(Lints::new(vec![DuplicatedTrailers]));
-
-        assert_eq!(
-            expected, actual,
-            "Expected the list of lint identifiers to be {:?}, instead got {:?}",
-            expected, actual
-        )
-    }
-
-    #[test]
-    fn enabled_pivotal_tracker_id() {
-        let mut strings = BTreeMap::new();
-        strings.insert("pb.lint.pivotal-tracker-id-missing".into(), "true".into());
-        let config = InMemory::new(&mut strings);
-
-        let actual = get_lint_configuration(&config);
-        let expected = Ok(Lints::new(vec![
-            DuplicatedTrailers,
-            PivotalTrackerIdMissing,
-        ]));
-
-        assert_eq!(
-            expected, actual,
-            "Expected the list of lint identifiers to be {:?}, instead got {:?}",
-            expected, actual
-        )
-    }
-
-    #[test]
-    fn enabled_jira_issue_key_missing() {
-        let mut strings = BTreeMap::new();
-        strings.insert("pb.lint.jira-issue-key-missing".into(), "true".into());
-        let config = InMemory::new(&mut strings);
-
-        let actual = get_lint_configuration(&config);
-        let expected = Ok(Lints::new(vec![DuplicatedTrailers, JiraIssueKeyMissing]));
-
-        assert_eq!(
-            expected, actual,
-            "Expected the list of lint identifiers to be {:?}, instead got {:?}",
-            expected, actual
-        )
-    }
-
-    #[test]
-    fn disabled_jira_issue_key_missing() {
-        let mut strings = BTreeMap::new();
-        strings.insert("pb.lint.jira-issue-key-missing".into(), "false".into());
-        let config = InMemory::new(&mut strings);
-
-        let actual = get_lint_configuration(&config);
-        let expected = Ok(Lints::new(vec![DuplicatedTrailers]));
-
-        assert_eq!(
-            expected, actual,
-            "Expected the list of lint identifiers to be {:?}, instead got {:?}",
-            expected, actual
-        )
-    }
-}
 
 #[cfg(test)]
 mod tests_can_enable_lints_via_a_command {
